# docs/swagger.yaml
basePath: /api/v1
definitions:
  docs.AlternativesResponse:
    properties:
      alternatives:
        items:
          $ref: '#/definitions/docs.FoodResponse'
        type: array
      food_id:
        type: string
      limit:
        type: integer
    type: object
  docs.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
...
  docs.FoodResponse:
    properties:
      calories:
        type: number
      carbohydrates:
        type: number
      category:
        type: string
      fat:
        type: number
      fiber:
        type: number
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      protein:
        type: number
      sodium:
        type: number
      sugar:
        type: number
    type: object
...
  docs.RecommendationRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      limit:
        type: integer
      max_calories:
        type: integer
      max_fat:
        type: number
      min_protein:
        type: number
      offset:
        type: integer
      profile_id:
        type: string
    type: object
  docs.RecommendationResponse:
    properties:
      applied_rules:
        items:
          type: string
        type: array
      pagination:
        properties:
          limit:
            type: integer
          offset:
            type: integer
        type: object
      recommendations:
        items:
          $ref: '#/definitions/docs.FoodResponse'
        type: array
      total_count:
        type: integer
    type: object
  docs.ReferenceItem:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  docs.Response:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nutrimatch.com
    name: API Support
    url: http://www.nutrimatch.com/support
  description: A personalized nutrition recommendation system built on the OpenNutrition
    dataset
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NutriMatch API
  version: "1.0"
paths:
...
  /api/v1/recommendations:
    get:
      consumes:
      - application/json
      description: Get personalized food recommendations for the authenticated user
      parameters:
      - default: 10
        description: Number of recommendations to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Profile ID to use for recommendations
        in: query
        name: profileId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.RecommendationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get food recommendations
      tags:
      - recommendations
  /api/v1/recommendations/alternatives/{foodId}:
    get:
      consumes:
      - application/json
      description: Get alternative food suggestions for a specific food item
      parameters:
      - description: Food ID
        in: path
        name: foodId
        required: true
        type: string
      - default: 5
        description: Number of alternatives to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.AlternativesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get food alternatives
      tags:
      - recommendations
  /api/v1/recommendations/filter:
    post:
      consumes:
      - application/json
      description: Filter food recommendations based on specific criteria
      parameters:
      - description: Filter criteria
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/docs.RecommendationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.RecommendationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Filter food recommendations
      tags:
      - recommendations
  /api/v1/reference/allergens:
    get:
      consumes:
      - application/json
      description: Get a list of all allergens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/docs.ReferenceItem'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Get allergens
      tags:
      - reference
  /api/v1/reference/dietary-patterns:
    get:
      consumes:
      - application/json
      description: Get a list of all dietary patterns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/docs.ReferenceItem'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Get dietary patterns
      tags:
      - reference
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"